+++ src/pages/common/DatasetWorkspace/plugins/text-letter-count/TextLetterCountView.tsx
@@ -53,7 +53,7 @@ const TextLetterCountView: FunctionComponent<DatasetPluginProps> = ({
         isRefreshing={isRefreshing}
         onSubmit={submitJob}
         onRefresh={fetchJobStatus}
-        submitButtonLabel="Analyze Letter Frequencies"
+        jobLabel="Analyze Letter Frequencies"
       />
     );
   }